+++ app/src/main/java/woowacourse/shopping/domain/usecase/cart/RemoveFromCartUseCase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain.usecase
+package woowacourse.shopping.domain.usecase.cart
 
 import woowacourse.shopping.domain.model.CartProduct
 import woowacourse.shopping.domain.repository.CartProductRepository
