+++ app/src/main/java/woowacourse/shopping/view/cart/select/CartProductSelectViewModelFactory.kt
@@ -2,14 +2,14 @@ package woowacourse.shopping.view.cart.select
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.domain.usecase.GetPagedCartProductsUseCase
-import woowacourse.shopping.domain.usecase.RemoveFromCartUseCase
-import woowacourse.shopping.domain.usecase.UpdateQuantityUseCase
+import woowacourse.shopping.domain.usecase.cart.GetPagedCartProductsUseCase
+import woowacourse.shopping.domain.usecase.cart.RemoveFromCartUseCase
+import woowacourse.shopping.domain.usecase.cart.UpdateCartQuantityUseCase
 
 class CartProductSelectViewModelFactory(
     private val getPagedCartProductsUseCase: GetPagedCartProductsUseCase,
     private val removeFromCartUseCase: RemoveFromCartUseCase,
-    private val updateQuantityUseCase: UpdateQuantityUseCase,
+    private val updateCartQuantityUseCase: UpdateCartQuantityUseCase,
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
@@ -17,7 +17,7 @@ class CartProductSelectViewModelFactory(
             return CartProductSelectViewModel(
                 getPagedCartProductsUseCase,
                 removeFromCartUseCase,
-                updateQuantityUseCase,
+                updateCartQuantityUseCase,
             ) as T
         }
         throw IllegalArgumentException()
