+++ app/src/main/java/woowacourse/shopping/view/product/catalog/ProductCatalogViewModelFactory.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.domain.repository.RecentProductRepository
-import woowacourse.shopping.domain.usecase.AddToCartUseCase
-import woowacourse.shopping.domain.usecase.GetPagedCartProductsUseCase
-import woowacourse.shopping.domain.usecase.GetTotalCartProductQuantityUseCase
-import woowacourse.shopping.domain.usecase.UpdateQuantityUseCase
+import woowacourse.shopping.domain.usecase.cart.AddToCartUseCase
+import woowacourse.shopping.domain.usecase.cart.GetPagedCartProductsUseCase
+import woowacourse.shopping.domain.usecase.cart.GetTotalCartProductQuantityUseCase
+import woowacourse.shopping.domain.usecase.cart.UpdateCartQuantityUseCase
 
 class ProductCatalogViewModelFactory(
     private val productRepository: ProductRepository,
@@ -15,7 +15,7 @@ class ProductCatalogViewModelFactory(
     private val getPagedCartProductsUseCase: GetPagedCartProductsUseCase,
     private val getTotalCartProductQuantityUseCase: GetTotalCartProductQuantityUseCase,
     private val addToCartUseCase: AddToCartUseCase,
-    private val updateQuantityUseCase: UpdateQuantityUseCase,
+    private val updateCartQuantityUseCase: UpdateCartQuantityUseCase,
 ) : ViewModelProvider.Factory {
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
@@ -26,7 +26,7 @@ class ProductCatalogViewModelFactory(
                 getPagedCartProductsUseCase,
                 getTotalCartProductQuantityUseCase,
                 addToCartUseCase,
-                updateQuantityUseCase,
+                updateCartQuantityUseCase,
             ) as T
         }
         throw IllegalArgumentException()
