+++ app/src/main/java/woowacourse/shopping/view/product/detail/ProductDetailActivity.kt
@@ -33,7 +33,7 @@ class ProductDetailActivity : AppCompatActivity() {
                     app.recentProductRepository,
                     app.getCartProductByProductIdUseCase,
                     app.addToCartUseCase,
-                    app.updateQuantityUseCase,
+                    app.updateCartQuantityUseCase,
                 ),
             )[ProductDetailViewModel::class.java]
         initBindings(product)
